Add testcases with cabal test
